CompetencyGraph/backend/app/pyproject.toml
[project]
name = "competency-graph"
version = "0.1.0"
description = "Competency Graph API"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "dependency-injector>=4.41.0",
    "SPARQLWrapper>=2.0.0",
    "orjson>=3.9.10",
    "asyncpg",
    "requests>=2.31.0",
    "PyJWT>=2.8.0",
    "redis>=5.0.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9"
]
